pyproject.toml
[project]
name = "rule-based-engine"
version = "0.1.0"
description = "AI Dispatch Engine Temp Mock Phase 1"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.103.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.23.0",
    "pandas",
    "scikit-learn",
    "lightgbm",
    "jupyter",
    "python-dotenv"
]
